# encoding/blockchain: add Reader type