… encoding/blockchain/bytereader_test.go