@jbowens 2017-04-20
> You don't have to calculate the value of n, copy will do it for you. Also, it's okay to return io.EOF along with some bytes in the last read. So this method can be condensed a little.

Noice, thanks. Updated