@jbowens 2017-04-20
just added some quickcheck tests for the encoding/blockchain package too. could also do that in a separate PR.